move getFoundationModelDeploymentId to ai api
