+++ packages/ai-api/src/utils/deployment-resolver.ts
@@ -4,7 +4,11 @@ import {
   DeploymentApi
 } from '../client/AI_CORE_API/index.js';
 import { deploymentCache } from './deployment-cache.js';
-import { extractModel, type FoundationModel } from './model.js';
+import {
+  extractModel,
+  translateToFoundationModel,
+  type FoundationModel
+} from './model.js';
 import type { HttpDestinationOrFetchOptions } from '@sap-cloud-sdk/connectivity';
 
 /**
@@ -198,4 +202,30 @@ export async function getAllDeployments(
   } catch (error: any) {
     throw new ErrorWithCause('Failed to fetch the list of deployments.', error);
   }
+}
+
+/**
+ * Get the deployment ID for a foundation model scenario.
+ * @param modelDeployment - This configuration is used to retrieve a deployment. Depending on the configuration use either the given deployment ID or the model name to retrieve matching deployments. If model and deployment ID are given, the model is verified against the deployment.
+ * @param executableId - The scenario ID.
+ * @param destination - The destination to use for the request.
+ * @returns The ID of the deployment, if found.
+ * @internal
+ */
+export async function getFoundationModelDeploymentId(
+  modelDeployment: ModelDeployment,
+  executableId: string,
+  destination?: HttpDestinationOrFetchOptions
+): Promise<string> {
+  if (isDeploymentIdConfig(modelDeployment)) {
+    return modelDeployment.deploymentId;
+  }
+
+  return resolveDeploymentId({
+    scenarioId: 'foundation-models',
+    executableId,
+    model: translateToFoundationModel(modelDeployment),
+    resourceGroup: getResourceGroup(modelDeployment),
+    destination
+  });
 }
